# Shatter letter delivery tasks