# Before

When delivering letters to DVLA we are bottlenecked by the speed at which we can create SQS messages. This is a problem because we have one task which delivers all the letters, and it creates one SQS message at a time. This task runs for a long time, and we can’t deploy while it’s running.

<img width="1298" alt="image" src="https://github.com/user-attachments/assets/d1ead061-e1be-4eb1-9d0e-20c816bb7203" />

# After 

This change makes it so that one new task (and SQS message) is created for each 100 letters. Each of these new tasks – which can be picked up asynchronously – then create individual tasks for each of its 100 letters.

The overall time taken will probably be about the same, but the maximum time taken by any one task should be greatly reduced.

<img width="1298" alt="image" src="https://github.com/user-attachments/assets/a3223d82-c232-4dac-9500-fcda46083c1f" />

# Estimated increase in performance

  | Number of  letters<sup>1</sup>  | Database time | Time talking to SQS | Total task time
-- | -- | -- | -- | -- 
**Before** | 166,936  | 11 seconds<sup>2</sup> | 29 minutes 15 seconds | 29 minutes 26 seconds<sup>1</sup>
**After** | 166,936  |  11 seconds | 17 seconds | 28 seconds 💥 

1. Taken from last night’s print run
2. Estimated by running the same query manually against production and scaling up based on number of results

***


This is a similar approach taken by `shatter_jobs` in https://github.com/alphagov/notifications-api/pull/4284