@quis 2025-02-27
Date | Number of letters | Time taken
---|---|---
25th February | 166,936 | 29 minutes 27 Seconds
26th February | 167,764 | 23 seconds

78&times; faster